server/requirements.txt

# Base Django requirements
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
drf-nested-routers==0.93.4

# Database connectors
mysqlclient==2.1.1
# SQLite3 est inclus dans la bibliothèque standard Python

# Image handling
Pillow==9.4.0

# Authentication
PyJWT==2.6.0

# Utils
python-dotenv==1.0.0
requests==2.28.2

